+++ scilpy/gradients/utils.py
